+++ package/router/src/trigger-click.ts
@@ -1,4 +1,5 @@
-import {logger, routerChangeSignal} from './core';
+import {logger} from './core';
+import {routeChangeSignal} from './signal';
 
 let _enabled = false;
 
@@ -31,23 +32,19 @@ export const clickTrigger = {
     }
 
     // find the <a> element that the click is at (or within)
-    let anchor = event
+    const anchor = event
         .composedPath()
-        .filter(
-            (eventTarget) => (eventTarget as HTMLElement).tagName.toLowerCase() === 'a',
-        )[0] as HTMLAnchorElement | null;
-
-    while (anchor != null && anchor.tagName.toLowerCase() !== 'a') {
-      anchor = anchor.parentNode as HTMLAnchorElement | null;
-    }
+        .find(
+            (target) => (target as HTMLElement)?.tagName?.toLowerCase() === 'a',
+        ) as HTMLAnchorElement | undefined;
 
     if (
       // ignore if the anchor is not found.
       anchor == null ||
       // ignore if the anchor is not an <a> element.
-      anchor.tagName.toLowerCase() !== 'a' ||
+      anchor.tagName?.toLowerCase() !== 'a' ||
       // ignore if the <a> element has a non-default target.
-      (anchor.target !== '' && anchor.target.toLowerCase() !== '_self' ) ||
+      (anchor.target !== '' && anchor.target?.toLowerCase() !== '_self' ) ||
       // ignore if the <a> element has a download attribute.
       anchor.hasAttribute('download') ||
       // ignore if the <a> element has a rel attribute.
@@ -71,7 +68,7 @@ export const clickTrigger = {
 
     // if none of the above, convert the click into a navigation signal.
     const {pathname, search, hash} = anchor;
-    routerChangeSignal.request({pathname, search, hash});
+    routeChangeSignal.request({pathname, search, hash});
     // for a click event, the scroll is reset to the top position.
     if (event.type === 'click') {
       window.scrollTo(0, 0);