+++ package/router/src/type.ts
@@ -1,13 +1,3 @@
-declare global {
-  interface AlwatrSignals {
-    'router-change': Route;
-  }
-
-  interface AlwatrRequestSignals {
-    'router-change': RequestRouteParam;
-  }
-}
-
 export type ParamList = Record<string, string | number | boolean>;
 
 // @TODO: description
@@ -16,7 +6,7 @@ export interface Route
   // href: https://example.com/product/100/book?cart=1&color=white#description
   sectionList: Array<string | number | boolean>; // [product, 100, book]
   queryParamList: ParamList; // {cart: 1, color: 'white'}
-  hash: string; // '#header'
+  hash: string; // '#description'
 }
 
 // @TODO: description
@@ -54,4 +44,9 @@ export interface InitOptions {
   popstateTrigger?: boolean;
 }
 
-
+export interface RoutesConfig {
+  map: (route: Route) => string | undefined;
+  list: Record<string, {
+    render: (route: Route) => unknown;
+  }>;
+}